changed names to be more descriptive
